+++ CHANGELOG.md
@@ -7,8 +7,14 @@ Versioning].
 
 ## [Unreleased]
 
+### Features
+
 - Run on STDIN. ([#119])
 - Process input files in parallel. ([#132])
+
+### Miscellaneous
+
+- Log when a mapping is overwritten. ([#134])
 
 ## [0.6.0-beta] - 2020-08-18
 
@@ -151,3 +157,4 @@ Versioning].
 [#118]: https://github.com/ericcornelissen/wordrow/pull/118
 [#119]: https://github.com/ericcornelissen/wordrow/pull/119
 [#132]: https://github.com/ericcornelissen/wordrow/pull/132
+[#134]: https://github.com/ericcornelissen/wordrow/pull/134